토큰 검증 로직 유틸 함수로 분리
